chore(patch): Removes unified param from diff.
This fixes an issue where some patches generated by ./create-patch.sh cannot be applied back to the repo.

Applying the patch back to the repo isn't a requirement for grading - so candidates can safely submit patches generated by the older version of the script.

Thanks to @MirisWisdom for reporting and helping reproduce this.